Fix the 1000x0 issue.
The cause: WAD files are cached in RAM to speed up random accesses (these are common in the engine). DVBBufferedImage is used as a container as it adds nothing to the memory limit of 8MB.
On startup DooM calls access() (which is implemented as 'fopen() -- check for NULL -- fclose()') on several known filenames to determine the gamemode. For nonexistent files File.length() returns 0, so the caching code would try to allocate a 1000x0 DVBBufferedImage to hold the contents. Fixed by adding 1 to the height.